+++ src/main/java/backend/wal/notification/adapter/FirebaseMessagingAdapter.java
@@ -1,6 +1,6 @@
 package backend.wal.notification.adapter;
 
-import backend.wal.notification.application.port.FirebaseMessagingPort;
+import backend.wal.notification.application.port.out.FirebaseMessagingPort;
 import com.google.api.core.ApiFuture;
 import com.google.api.core.ApiFutureCallback;
 import com.google.api.core.ApiFutures;
