+++ src/main/java/backend/wal/notification/application/port/in/NotificationUseCase.java
@@ -1,4 +1,4 @@
-package backend.wal.notification.application.port;
+package backend.wal.notification.application.port.in;
 
 public interface NotificationUseCase {
