+++ src/main/java/backend/wal/notification/application/port/in/RegisterFcmTokenUseCase.java
@@ -1,6 +1,4 @@
-package backend.wal.notification.application.port;
-
-import backend.wal.auth.application.port.dto.InitFcmRequestDto;
+package backend.wal.notification.application.port.in;
 
 public interface RegisterFcmTokenUseCase {
